Module type Time.S_kernel

module Time : sig ... end

include module type of struct include Time end

type underlying = Time.underlying

type t = private underlying

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

module Span = Time.Span

module Ofday = Time.Ofday

module Replace_polymorphic_compare = Time.Replace_polymorphic_compare

include Core.Comparable.Polymorphic_compare with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

include Core.Robustly_comparable.S with type t := t

val (>=.) : t -> t -> bool

val (<=.) : t -> t -> bool

val (=.) : t -> t -> bool

val (>.) : t -> t -> bool

val (<.) : t -> t -> bool

val (<>.) : t -> t -> bool

val robustly_compare : t -> t -> int

val next : t -> t

next t returns the next t (forwards in time)

val prev : t -> t

prev t returns the previous t (backwards in time)

val to_span_since_epoch : t -> Span.t

val of_span_since_epoch : Span.t -> t

include Core.Comparable.S_common with type t := t and module Replace_polymorphic_compare := Replace_polymorphic_compare

include Base.Comparable.S with type t := t

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness = Time.comparator_witness

val comparator : ( t, comparator_witness ) Base.Comparator.comparator

val validate_lbound : min:t Core.Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Core.Maybe_bound.t -> t Validate.check

val validate_bound : min:t Core.Maybe_bound.t -> max:t Core.Maybe_bound.t -> t Validate.check

module Date_and_ofday = Time.Date_and_ofday

Equivalent to a Date.t and an Ofday.t with no time zone. A Date_and_ofday.t does not correspond to a single, unambiguous point in time.

val next_multiple : ?can_equal_after:Base.Bool.t -> base:t -> after:t -> interval:Span.t -> Base.Unit.t -> t

next_multiple ~base ~after ~interval returns the smallest time of the form:

time = base + k * interval

where k >= 0 and time > after. It is an error if interval <= 0.

Supplying ~can_equal_after:true allows the result to satisfy time >= after.

val prev_multiple : ?can_equal_before:Base.Bool.t -> base:t -> before:t -> interval:Span.t -> Base.Unit.t -> t

prev_multiple ~base ~before ~interval returns the largest time of the form:

time = base + k * interval

where k >= 0 and time < before. It is an error if interval <= 0.

Supplying ~can_equal_before:true allows the result to satisfy time <= before.

val now : Base.Unit.t -> t

now () returns a t representing the current time

module Zone : sig ... end

Basic operations on times

val add : t -> Span.t -> t

add t s adds the span s to time t and returns the resulting time.

NOTE: adding spans as a means of adding days is not accurate, and may run into trouble due to shifts in daylight savings time, float arithmetic issues, and leap seconds. See the comment at the top of Zone.mli for a more complete discussion of some of the issues of time-keeping. For spans that cross date boundaries, use date functions instead.

val sub : t -> Span.t -> t

sub t s subtracts the span s from time t and returns the resulting time. See important note for add.

val diff : t -> t -> Span.t

diff t1 t2 returns time t1 minus time t2.

val abs_diff : t -> t -> Span.t

abs_diff t1 t2 returns the absolute span of time t1 minus time t2.

include Core.Quickcheck.S_range with type t := t

include Core.Quickcheck_intf.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val gen_incl : t -> t -> t Base_quickcheck.Generator.t

gen_incl lower_bound upper_bound produces values between lower_bound and upper_bound, inclusive. It uses an ad hoc distribution that stresses boundary conditions more often than a uniform distribution, while still able to produce any value in the range. Raises if lower_bound > upper_bound.

val gen_uniform_incl : t -> t -> t Base_quickcheck.Generator.t

gen_uniform_incl lower_bound upper_bound produces a generator for values uniformly distributed between lower_bound and upper_bound, inclusive. Raises if lower_bound > upper_bound.

Comparisons

val is_earlier : t -> than:t -> Base.Bool.t

val is_later : t -> than:t -> Base.Bool.t

Conversions

val of_date_ofday : zone:Core__Zone.t -> Core__Date0.t -> Ofday.t -> t

val of_date_ofday_precise : Core__Date0.t -> Ofday.t -> zone:Core__Zone.t -> [ `Once of t | `Twice of t * t | `Never of t ]

Because timezone offsets change throughout the year (clocks go forward or back) some local times can occur twice or not at all. In the case that they occur twice, this function gives `Twice with both occurrences in order; if they do not occur at all, this function gives `Never with the time at which the local clock skips over the desired time of day.

Note that this is really only intended to work with DST transitions and not unusual or dramatic changes, like the calendar change in 1752 (run "cal 9 1752" in a shell to see). In particular it makes the assumption that midnight of each day is unambiguous.

Most callers should use of_date_ofday rather than this function. In the `Twice and `Never cases, of_date_ofday will return reasonable times for most uses.

val to_date_ofday : t -> zone:Core__Zone.t -> Core__Date0.t * Ofday.t

val to_date_ofday_precise : t -> zone:Core__Zone.t -> Core__Date0.t * Ofday.t * [ `Only | `Also_at of t | `Also_skipped of Core__Date0.t * Ofday.t ]

Always returns the Date.t * Ofday.t that to_date_ofday would have returned, and in addition returns a variant indicating whether the time is associated with a time zone transition.

      - `Only         -> there is a one-to-one mapping between [t]'s and
                         [Date.t * Ofday.t] pairs
      - `Also_at      -> there is another [t] that maps to the same [Date.t * Ofday.t]
                         (this date/time pair happened twice because the clock fell back)
      - `Also_skipped -> there is another [Date.t * Ofday.t] pair that never happened (due
                         to a jump forward) that [of_date_ofday] would map to the same
                         [t].

val to_date : t -> zone:Core__Zone.t -> Core__Date0.t

val to_ofday : t -> zone:Core__Zone.t -> Ofday.t

val reset_date_cache : Base.Unit.t -> Base.Unit.t

For performance testing only; reset_date_cache () resets an internal cache used to speed up to_date and related functions when called repeatedly on times that fall within the same day.

Unlike Time_ns, this module purposely omits max_value and min_value: 1. They produce unintuitive corner cases because most people's mental models of time do not include +/- infinity as concrete values 2. In practice, when people ask for these values, it is for questionable uses, e.g., as null values to use in place of explicit options.

val epoch : t

midnight, Jan 1, 1970 in UTC

val convert : from_tz:Core__Zone.t -> to_tz:Core__Zone.t -> Core__Date0.t -> Ofday.t -> Core__Date0.t * Ofday.t

It's unspecified what happens if the given date/ofday/zone correspond to more than one date/ofday pair in the other zone.

val utc_offset : t -> zone:Core__Zone.t -> Span.t

Other string conversions

The {to,of}_string functions in Time convert to UTC time, because a local time zone is not necessarily available. They are generous in what they will read in.

include Base.Stringable.S with type t := t

val to_filename_string : t -> zone:Core__Zone.t -> Base.String.t

to_filename_string t ~zone converts t to string with format YYYY-MM-DD_HH-MM-SS.mmm which is suitable for using in filenames.

val of_filename_string : Base.String.t -> zone:Core__Zone.t -> t

of_filename_string s ~zone converts s that has format YYYY-MM-DD_HH-MM-SS.mmm into time.

val to_string_abs : t -> zone:Core__Zone.t -> Base.String.t

to_string_abs ~zone t is the same as to_string t except that it uses the given time zone.

val to_string_abs_trimmed : t -> zone:Core__Zone.t -> Base.String.t

to_string_abs_trimmed is the same as to_string_abs, but drops trailing seconds and milliseconds if they are 0.

val to_string_abs_parts : t -> zone:Core__Zone.t -> Base.String.t Base.List.t

val to_string_trimmed : t -> zone:Core__Zone.t -> Base.String.t

Same as to_string_abs_trimmed, except it leaves off the timezone, so won't reliably round trip.

val to_sec_string : t -> zone:Core__Zone.t -> Base.String.t

Same as to_string_abs, but without milliseconds and the timezone

val to_sec_string_with_zone : t -> zone:Core__Zone.t -> Base.String.t

Same as to_sec_string but includes timezone

val of_localized_string : zone:Core__Zone.t -> Base.String.t -> t

of_localized_string ~zone str read in the given string assuming that it represents a time in zone and return the appropriate Time.t

val of_string_gen : default_zone:( Base.Unit.t -> Core__Zone.t ) -> find_zone:( Base.String.t -> Core__Zone.t ) -> Base.String.t -> t

of_string_gen ~default_zone ~find_zone s attempts to parse s as a t, calling out to default_zone and find_zone as needed.

val to_string_iso8601_basic : t -> zone:Core__Zone.t -> Base.String.t

to_string_iso8601_basic return a string representation of the following form: %Y-%m-%dT%H:%M:%S.%s%Z e.g. to_string_iso8601_basic ~zone:Time.Zone.utc epoch = "1970-01-01T00:00:00.000000Z"

val occurrence : [ `First_after_or_at | `Last_before_or_at ] -> t -> ofday:Ofday.t -> zone:Core__Zone.t -> t

occurrence side time ~ofday ~zone returns a Time.t that is the occurrence of ofday (in the given zone) that is the latest occurrence (<=) time or the earliest occurrence (>=) time, according to side.

NOTE: If the given time converted to wall clock time in the given zone is equal to ofday then the t returned will be equal to the t given.

val of_string : Base.String.t -> t

val of_string_with_utc_offset : Base.String.t -> t

of_string_with_utc_offset requires its input to have an explicit UTC offset, e.g. 2000-01-01 12:34:56.789012-23, or use the UTC zone, "Z", e.g. 2000-01-01 12:34:56.789012Z.

val to_string : t -> Base.String.t

val to_string_utc : t -> Base.String.t

to_string_utc generates a time string with the UTC zone, "Z", e.g. 2000-01-01 12:34:56.789012Z.