Module Deferred.Result

include Core.Monad.S2 with type ('a, 'b) t = ( 'a, 'b ) Core.Result.t Deferred.t

type ('a, 'b) t = ( 'a, 'b ) Core.Result.t Deferred.t

val (>>=) : ( 'a, 'e ) t -> ( 'a -> ( 'b, 'e ) t ) -> ( 'b, 'e ) t

val (>>|) : ( 'a, 'e ) t -> ( 'a -> 'b ) -> ( 'b, 'e ) t

module Let_syntax : sig ... end

module Monad_infix : sig ... end

Same as Infix, except the monad type has two arguments. The second is always just passed through.

val bind : ( 'a, 'e ) t -> f:( 'a -> ( 'b, 'e ) t ) -> ( 'b, 'e ) t

val return : 'a -> ( 'a, _ ) t

val map : ( 'a, 'e ) t -> f:( 'a -> 'b ) -> ( 'b, 'e ) t

val join : ( ( 'a, 'e ) t, 'e ) t -> ( 'a, 'e ) t

val ignore_m : ( _, 'e ) t -> ( unit, 'e ) t

val all : ( 'a, 'e ) t list -> ( 'a list, 'e ) t

val all_unit : ( unit, 'e ) t list -> ( unit, 'e ) t

val fail : 'err -> ( _, 'err ) t

val failf : ( 'a, unit, string, ( _, string ) t ) Core.format4 -> 'a

e.g., failf "Couldn't find bloogle %s" (Bloogle.to_string b).

val map_error : ( 'ok, 'error1 ) t -> f:( 'error1 -> 'error2 ) -> ( 'ok, 'error2 ) t

val combine : ( 'ok1, 'err ) t -> ( 'ok2, 'err ) t -> ok:( 'ok1 -> 'ok2 -> 'ok3 ) -> err:( 'err -> 'err -> 'err ) -> ( 'ok3, 'err ) t

combine waits on both inputs and combines their results using Result.combine.