Module Async_kernel.Monad_sequence

Monad_sequence.S is a generic interface specifying functions that deal with a container and a monad. It is specialized to the Deferred monad and used with various containers in modules Deferred.Array, Deferred.List, Deferred.Queue, and Deferred.Sequence. The Monad_sequence.how type specifies the parallelism of container iterators.

type how = [

| `Parallel	(*
like `Max_concurrent_jobs Int.max_value
*)

| `Sequential	(*
`Sequential is often but not always the same as `Max_concurrent_jobs 1 (for example, they differ in the Or_error monad).
*)

| `Max_concurrent_jobs of int

]

val sexp_of_how : how -> Sexplib0.Sexp.t

module type S = sig ... end