Module Monitor.Monitor_exn

type t

val extract_exn : t -> exn

Extract the exception as it was originally received by an async monitor.

val backtrace : t -> Core.Backtrace.t option

Extract the backtrace originally captured at the point the exception was received by an async monitor.