Module Synchronous_time_source.Time_ns

include module type of struct include Core.Core_private.Time_ns_alternate_sexp end

include module type of struct include Core.Time_ns end

include Bin_prot.Binable.S__local with type t := Core.Int63.t

include Bin_prot.Binable.S_only_functions__local with type t := Core.Int63.t

include sig ... end

include Ppx_compare_lib.Comparable.S with type t := Core.Int63.t

include Ppx_compare_lib.Comparable.S__local with type t := Core.Int63.t

include Ppx_compare_lib.Equal.S with type t := Core.Int63.t

include Ppx_compare_lib.Equal.S__local with type t := Core.Int63.t

val globalize : Core.Int63.t @ local -> Core.Int63.t

include Ppx_hash_lib.Hashable.S_any with type t := Core.Int63.t

include Sexplib0.Sexpable.S__stack with type t := Core.Int63.t

include Sexplib0.Sexpable.Of_sexp with type t := Core.Int63.t

include Sexplib0.Sexpable.Sexp_of__stack with type t := Core.Int63.t

include Typerep_lib.Typerepable.S with type t := Core.Int63.t

val typerep_of_t : Core.Int63.t Typerep_lib.Std_internal.Typerep.t @@ portable

val typename_of_t : Core.Int63.t Typerep_lib.Typename.t @@ portable

module Span = Core.Core_private.Time_ns_alternate_sexp.Span

module Ofday = Core.Core_private.Time_ns_alternate_sexp.Ofday

include Base.Comparisons.S_with_zero_alloc__local with type t := Core.Int63.t

include Base.Comparisons.Infix_with_zero_alloc with type t := Core.Int63.t

include Diffable.S_atomic with type t := Core.Int63.t

module Zone = Core.Core_private.Time_ns_alternate_sexp.Zone

module Alternate_sexp = Core.Core_private.Time_ns_alternate_sexp.Alternate_sexp

Note that we expose a sexp format that is not the one exposed in Time_ns_unix. The sexp is a single atom rendered as with to_string_utc, except that all trailing zeros are trimmed, rather than trimming in groups of three.

module Option = Core.Core_private.Time_ns_alternate_sexp.Option

Option.t is like t option, except that the value is immediate. This module should mainly be used to avoid allocations.

include Core.Quickcheck.S_range with type t := Core.Int63.t

include Core.Quickcheck_intf.Definitions.S with type t := Core.Int63.t

val quickcheck_generator : Core.Int63.t Base_quickcheck.Generator.t

val quickcheck_observer : Core.Int63.t Base_quickcheck.Observer.t

val quickcheck_shrinker : Core.Int63.t Base_quickcheck.Shrinker.t

val gen_incl : Core.Int63.t -> Core.Int63.t -> Core.Int63.t Base_quickcheck.Generator.t

gen_incl lower_bound upper_bound produces values between lower_bound and upper_bound, inclusive. It uses an ad hoc distribution that stresses boundary conditions more often than a uniform distribution, while still able to produce any value in the range. Raises if lower_bound > upper_bound.

val gen_uniform_incl : Core.Int63.t -> Core.Int63.t -> Core.Int63.t Base_quickcheck.Generator.t

gen_uniform_incl lower_bound upper_bound produces a generator for values uniformly distributed between lower_bound and upper_bound, inclusive. Raises if lower_bound > upper_bound.

Comparisons

val is_earlier : Core.Int63.t -> than:Core.Int63.t -> Base.Bool.t

val is_later : Core.Int63.t -> than:Core.Int63.t -> Base.Bool.t

Conversions

val of_date_ofday : ?prefer:Core__.Zone_intf.Earlier_or_later.t -> zone:Core__.Zone.t -> Core__.Date0.t -> Ofday.t -> Core.Int63.t

Produces the absolute time of a date and time of day in this zone. When there are two occurrences, the result is determined by prefer.

val of_date_ofday_precise : Core__.Date0.t -> Ofday.t -> zone:Core__.Zone.t -> [ `Once of Core.Int63.t | `Twice of Core.Int63.t * Core.Int63.t | `Never of Core.Int63.t ]

Because timezone offsets change throughout the year (clocks go forward or back) some local times can occur twice or not at all. In the case that they occur twice, this function gives `Twice with both occurrences in order; if they do not occur at all, this function gives `Never with the time at which the local clock skips over the desired time of day.

Note that this is really only intended to work with DST transitions and not unusual or dramatic changes, like the calendar change in 1752 (run "cal 9 1752" in a shell to see). In particular it makes the assumption that midnight of each day is unambiguous.

Most callers should use of_date_ofday rather than this function. In the `Twice and `Never cases, of_date_ofday will return reasonable times for most uses.

val to_date_ofday : Core.Int63.t -> zone:Core__.Zone.t -> Core__.Date0.t * Ofday.t

val to_date_ofday_precise : Core.Int63.t -> zone:Core__.Zone.t -> Core__.Date0.t * Ofday.t * [ `Only | `Also_at of Core.Int63.t | `Also_skipped of Core__.Date0.t * Ofday.t ]

Always returns the Date.t * Ofday.t that to_date_ofday would have returned, and in addition returns a variant indicating whether the time is associated with a time zone transition.

- `Only         -> there is a one-to-one mapping between [t]'s and
                   [Date.t * Ofday.t] pairs
- `Also_at      -> there is another [t] that maps to the same [Date.t * Ofday.t]
                   (this date/time pair happened twice because the clock fell back)
- `Also_skipped -> there is another [Date.t * Ofday.t] pair that never happened (due
                   to a jump forward) that [of_date_ofday] would map to the same
                   [t].

val to_date : Core.Int63.t -> zone:Core__.Zone.t -> Core__.Date0.t

val to_ofday : Core.Int63.t -> zone:Core__.Zone.t -> Ofday.t

val reset_date_cache : Base.Unit.t -> Base.Unit.t

For performance testing only; reset_date_cache () resets an internal cache used to speed up to_date and related functions when called repeatedly on times that fall within the same day.

Unlike Time_ns, this module purposely omits max_value and min_value: 1. They produce unintuitive corner cases because most people's mental models of time do not include +/- infinity as concrete values 2. In practice, when people ask for these values, it is for questionable uses, e.g., as null values to use in place of explicit options.

val convert : from_tz:Core__.Zone.t -> to_tz:Core__.Zone.t -> Core__.Date0.t -> Ofday.t -> Core__.Date0.t * Ofday.t

It's unspecified what happens if the given date/ofday/zone correspond to more than one date/ofday pair in the other zone.

val utc_offset : Core.Int63.t -> zone:Core__.Zone.t -> Span.t

Other string conversions

to_string uses the local time zone. of_string defaults to the local time zone if not specified, and is generous in what it will read in.

include Core.Interfaces.Stringable with type t := Core.Int63.t

val to_filename_string : Core.Int63.t -> zone:Core__.Zone.t -> Base.String.t

to_filename_string t ~zone converts t to string with format YYYY-MM-DD_HH-MM-SS.mmm which is suitable for using in filenames.

val of_filename_string : Base.String.t -> zone:Core__.Zone.t -> Core.Int63.t

of_filename_string s ~zone converts s that has format YYYY-MM-DD_HH-MM-SS.mmm into time.

val to_string_abs : Core.Int63.t -> zone:Core__.Zone.t -> Base.String.t

to_string_abs ~zone t is the same as to_string t except that it uses the given time zone.

val to_string_abs_trimmed : Core.Int63.t -> zone:Core__.Zone.t -> Base.String.t

to_string_abs_trimmed is the same as to_string_abs, but drops trailing seconds and milliseconds if they are 0.

val to_string_abs_parts : Core.Int63.t -> zone:Core__.Zone.t -> Base.String.t Base.List.t

val to_string_trimmed : Core.Int63.t -> zone:Core__.Zone.t -> Base.String.t

Same as to_string_abs_trimmed, except it leaves off the timezone, so won't reliably round trip.

val to_sec_string : Core.Int63.t -> zone:Core__.Zone.t -> Base.String.t

Same as to_string_abs, but without milliseconds and the timezone. May raise if zone offsets move the apparent value beyond min_value_representable and max_value_representable.

val to_sec_string_with_zone : Core.Int63.t -> zone:Core__.Zone.t -> Base.String.t

Same as to_sec_string but includes timezone

val of_localized_string : zone:Core__.Zone.t -> Base.String.t -> Core.Int63.t

of_localized_string ~zone str read in the given string assuming that it represents a time in zone and return the appropriate Time.t

val to_string_iso8601_basic : Core.Int63.t -> zone:Core__.Zone.t -> Base.String.t

to_string_iso8601_basic return a string representation of the following form: %Y-%m-%dT%H:%M:%S.%s%Z e.g. to_string_iso8601_basic ~zone:Time.Zone.utc epoch = "1970-01-01T00:00:00.000000Z"

val occurrence : [ `First_after_or_at | `Last_before_or_at ] -> Core.Int63.t -> ofday:Ofday.t -> zone:Core__.Zone.t -> Core.Int63.t

occurrence side time ~ofday ~zone returns a Time.t that is the occurrence of ofday (in the given zone) that is the latest occurrence (<=) time or the earliest occurrence (>=) time, according to side.

NOTE: If the given time converted to wall clock time in the given zone is equal to ofday then the t returned will be equal to the t given.

These functions are identical to those in Time and get/set the same variable.

val get_sexp_zone : Base.Unit.t -> Zone.t @@ portable

val set_sexp_zone : Zone.t -> Base.Unit.t @@ portable

val t_of_sexp_abs : Core.Sexp.t -> Core.Int63.t @@ portable

t_of_sexp_abs sexp as t_of_sexp, but demands that sexp indicate the timezone the time is expressed in.

val sexp_of_t_abs : Core.Int63.t -> zone:Zone.t -> Core.Sexp.t @@ portable

val of_date_ofday_zoned : Core.Date.t -> Ofday.Zoned.t -> Core.Int63.t @@ portable

Conversion functions that involved Ofday.Zoned.t, exactly analogous to the conversion functions that involve Ofday.t

val to_date_ofday_zoned : Core.Int63.t -> zone:Zone.t -> Core.Date.t * Ofday.Zoned.t @@ portable

val to_ofday_zoned : Core.Int63.t -> zone:Zone.t -> Ofday.Zoned.t @@ portable

val to_string_fix_proto : [ `Utc | `Local ] -> Core.Int63.t -> Base.String.t @@ portable

val of_string_fix_proto : [ `Utc | `Local ] -> Base.String.t -> Core.Int63.t @@ portable

val of_string_abs : Base.String.t -> Core.Int63.t @@ portable

This is like of_string except that if the string doesn't specify the zone then it raises rather than assume the local timezone.

val of_string_gen : if_no_timezone: [ `Fail | `Local | `Use_this_one of Zone.t | `Use_this_one_lazy of Zone.t Core.Lazy.t | `Use_this_one_portable_lazy of Zone.t Core.Portable_lazy.t ] -> ?find_zone:(Base.String.t -> Zone.t) -> Base.String.t -> Core.Int63.t @@ portable

of_string_gen ~if_no_timezone ?find_zone s attempts to parse s to a t. If s doesn't supply a time zone if_no_timezone is consulted. find_zone is used to look up time zones by name.

val of_string_with_utc_offset : Base.String.t -> Core.Int63.t @@ portable

of_string_with_utc_offset requires its input to have an explicit UTC offset, e.g. 2000-01-01 12:34:56.789012-23, or use the UTC zone, "Z", e.g. 2000-01-01 12:34:56.789012Z.

val to_string_utc : Core.Int63.t -> Base.String.t @@ portable

to_string_utc generates a time string with the UTC zone, "Z", e.g. 2000-01-01 12:34:56.789012Z.

val to_string_iso8601_extended : ?precision:[ `sec | `ms | `us | `ns ] -> zone:Zone.t -> Core.Int63.t -> Base.String.t @@ portable

Returns the ISO 8601 extended format for a time (including the date) in the given time zone. The number of decimal places is determined by precision, default `ns.

For example:

  to_string_iso8601_extended epoch ~zone:Zone.utc ~precision:`us
  = "1970-01-01T00:00:00.000000Z"

val epoch : Core.Int63.t @@ portable

Unix epoch (1970-01-01 00:00:00 UTC)

val min_value_representable : Core.Int63.t @@ portable

The minimum representable time.

val max_value_representable : Core.Int63.t @@ portable

The maximum representable time.

val min_value_for_1us_rounding : Core.Int63.t @@ portable

The minimum time that rounds to a Time.t with microsecond precision.

val max_value_for_1us_rounding : Core.Int63.t @@ portable

The maximum time that rounds to a Time.t with microsecond precision.

val min_value : Core.Int63.t @@ portable

An alias for min_value_for_1us_rounding.

deprecated [since 2019-02] use [min_value_representable] or [min_value_for_1us_rounding] instead

val max_value : Core.Int63.t @@ portable

An alias for max_value_for_1us_rounding.

deprecated [since 2019-02] use [max_value_representable] or [max_value_for_1us_rounding] instead

val now : Base.Unit.t -> Core.Int63.t @@ portable

The current time.

val add : Core.Int63.t -> Span.t -> Core.Int63.t @@ portable

overflows silently

val add_saturating : Core.Int63.t -> Span.t -> Core.Int63.t @@ portable

As add; rather than over/underflowing, clamps the result to the closed interval between min_value_representable and max_value_representable.

val sub_saturating : Core.Int63.t -> Span.t -> Core.Int63.t @@ portable

As sub; rather than over/underflowing, clamps the result to the closed interval between min_value_representable and max_value_representable.

val sub : Core.Int63.t -> Span.t -> Core.Int63.t @@ portable

overflows silently

val next : Core.Int63.t -> Core.Int63.t @@ portable

overflows silently

val prev : Core.Int63.t -> Core.Int63.t @@ portable

overflows silently

val diff : Core.Int63.t -> Core.Int63.t -> Span.t @@ portable

diff t1 t2 returns time t1 minus time t2.

This operation overflows silently.

val abs_diff : Core.Int63.t -> Core.Int63.t -> Span.t @@ portable

abs_diff t1 t2 returns the absolute span of time t1 minus time t2.

This operation overflows silently.

val to_span_since_epoch : Core.Int63.t -> Span.t @@ portable

val of_span_since_epoch : Span.t -> Core.Int63.t @@ portable

val to_int63_ns_since_epoch : Core.Int63.t -> Core.Int63.t @@ portable

val of_int63_ns_since_epoch : Core.Int63.t -> Core.Int63.t @@ portable

val to_int_ns_since_epoch : Core.Int63.t -> Base.Int.t @@ portable

Will raise on 32-bit platforms. Consider to_int63_ns_since_epoch instead.

val of_int_ns_since_epoch : Base.Int.t -> Core.Int63.t @@ portable

val next_multiple : ?can_equal_after:Base.Bool.t @ local -> (base:Core.Int63.t -> (after:Core.Int63.t -> (interval:Span.t -> (Base.Unit.t -> Core.Int63.t) @ local) @ local) @ local) @ local @@ portable

next_multiple ~base ~after ~interval returns the smallest time of the form:

  time = base + (k * interval)

where k >= 0 and time > after. It is an error if interval <= 0.

Supplying ~can_equal_after:true allows the result to satisfy time >= after.

This function is useful for finding linear time intervals, like every 30 minutes or every 24 hours. This is different from rounding to apparent clock-face internvals, like "every hour at :00 and :30" or "every day at noon", because time zone and daylight savings transitions may cause linear intervals and apparent clock-face intervals to differ.

Time zone offsets (in the tzdata time zone database, at least) are expressed in seconds, so rounding to units of seconds or smaller is not affected by time zones. The round* functions below provide some straightforward cases. For other small units that evenly divide into a second, call with base = epoch, after as the time to round, and interval as the unit span you are rounding to.

val prev_multiple : ?can_equal_before:Base.Bool.t -> base:Core.Int63.t -> before:Core.Int63.t -> interval:Span.t -> Base.Unit.t -> Core.Int63.t @@ portable

prev_multiple ~base ~before ~interval returns the largest time of the form:

  time = base + (k * interval)

where k >= 0 and time < before. It is an error if interval <= 0.

Supplying ~can_equal_before:true allows the result to satisfy time <= before.

This function is useful for finding linear time intervals, like every 30 minutes or every 24 hours. This is different from rounding to apparent clock-face internvals, like "every hour at :00 and :30" or "every day at noon", because time zone and daylight savings transitions may cause linear intervals and apparent clock-face intervals to differ.

Time zone offsets (in the tzdata time zone database, at least) are expressed in seconds, so rounding to units of seconds or smaller is not affected by time zones. The round* functions below provide some straightforward cases. For other small units that evenly divide into a second, call with base = epoch, after as the time to round, and interval as the unit span you are rounding to.

val round_up_to_us : Core.Int63.t -> Core.Int63.t @@ portable

round_up_to_us t returns t rounded up to the next microsecond.

val round_up_to_ms : Core.Int63.t -> Core.Int63.t @@ portable

round_up_to_ms t returns t rounded up to the next millisecond.

val round_up_to_sec : Core.Int63.t -> Core.Int63.t @@ portable

round_up_to_sec t returns t rounded up to the next second.

val round_up_to : Core.Int63.t -> interval:Span.t -> Core.Int63.t @@ portable

round_up_to returns t rounded up to the next interval

val round_down_to_us : Core.Int63.t -> Core.Int63.t @@ portable

round_down_to_us t returns t rounded down to the previous microsecond.

val round_down_to_ms : Core.Int63.t -> Core.Int63.t @@ portable

round_down_to_ms t returns t rounded down to the previous millisecond.

val round_down_to_sec : Core.Int63.t -> Core.Int63.t @@ portable

round_down_to_sec t returns t rounded down to the previous second.

val round_down_to : Core.Int63.t -> interval:Span.t -> Core.Int63.t @@ portable

round_down_to returns t rounded down to the previous interval

val every : Span.t -> start:Core.Int63.t -> stop:Core.Int63.t -> Core.Int63.t Base.List.t Core.Or_error.t @@ portable

every interval ~start ~stop returns a sorted list of all ts that can be expressed as start + (i * interval) and satisfy t >= start && t <= stop.

Returns an Error if interval <= Span.zero || start > stop.

val random : ?state:Base.Random.State.t -> Base.Unit.t -> Core.Int63.t @@ portable

val of_time : Core.Time_float.t -> Core.Int63.t @@ portable

deprecated [since 2019-01] use [of_time_float_round_nearest] or [of_time_float_round_nearest_microsecond]

val to_time : Core.Int63.t -> Core.Time_float.t @@ portable

deprecated [since 2019-01] use [to_time_float_round_nearest] or [to_time_float_round_nearest_microsecond]

*_round_nearest vs *_round_nearest_microsecond: If you don't know that you need microsecond precision, use the *_round_nearest version. *_round_nearest_microsecond is for historical purposes.

val to_time_float_round_nearest : Core.Int63.t -> Core.Time_float.t @@ portable

val to_time_float_round_nearest_microsecond : Core.Int63.t -> Core.Time_float.t @@ portable

val of_time_float_round_nearest : Core.Time_float.t -> Core.Int63.t @@ portable

val of_time_float_round_nearest_microsecond : Core.Time_float.t -> Core.Int63.t @@ portable

module Utc = Core.Core_private.Time_ns_alternate_sexp.Utc

module O = Core.Core_private.Time_ns_alternate_sexp.O

String conversions use the local timezone by default. Sexp conversions use get_sexp_zone () by default, which can be overridden by calling set_sexp_zone. These default time zones are used when writing a time, and when reading a time with no explicit zone or UTC offset.

Sexps and strings display the date, ofday, and UTC offset of t relative to the appropriate time zone.

include Core.Identifiable.S__local__portable with type t := Core.Int63.t

include Bin_prot.Binable.S__local with type t := Core.Int63.t

include Bin_prot.Binable.S_only_functions__local with type t := Core.Int63.t

include sig ... end

include Ppx_hash_lib.Hashable.S_any with type t := Core.Int63.t

include Sexplib0.Sexpable.S with type t := Core.Int63.t

include Sexplib0.Sexpable.Of_sexp with type t := Core.Int63.t

include Sexplib0.Sexpable.Sexp_of with type t := Core.Int63.t

include Ppx_compare_lib.Comparable.S with type t := Core.Int63.t

include Ppx_compare_lib.Comparable.S__local with type t := Core.Int63.t

include Ppx_hash_lib.Hashable.S_any with type t := Core.Int63.t

include Base.Stringable.S with type t := Core.Int63.t

val of_string : string -> Core.Int63.t

val to_string : Core.Int63.t -> string

include Base.Pretty_printer.S with type t := Core.Int63.t

val pp : Base.Formatter.t -> Core.Int63.t -> unit

include Base.Comparable.S__local__portable with type t := Core.Int63.t

include Base.Comparisons.S__local with type t := Core.Int63.t

include Base.Comparisons.Infix with type t := Core.Int63.t

include Base.Comparator.S__portable with type t := Core.Int63.t

include Core.Comparator.S__portable with type t := Core.Int63.t with type comparator_witness := Core.Time_ns.comparator_witness

include Core.Hashable.S_binable with type t := Core.Int63.t

include Ppx_hash_lib.Hashable.S_any with type t := Core.Int63.t

val hashable : Core.Int63.t Base.Hashable.t

module Table = Core.Core_private.Time_ns_alternate_sexp.Table

module Hash_set = Core.Core_private.Time_ns_alternate_sexp.Hash_set

module Hash_queue = Core.Core_private.Time_ns_alternate_sexp.Hash_queue

Identifiable masks the comparison functions' @zero_alloc annotations

include Base.Comparisons.S_with_zero_alloc_strict__local with type t := Core.Int63.t

include Base.Comparisons.Infix_with_zero_alloc_strict with type t := Core.Int63.t

module Stable = Core.Core_private.Time_ns_alternate_sexp.Stable

val arg_type : Core.Int63.t Command.Arg_type.t @@ portable

val interruptible_pause : [ `Use_Time_ns_unix ] @@ portable

deprecated [since 2021-03] Use [Time_ns_unix]

val pause : [ `Use_Time_ns_unix ] @@ portable

deprecated [since 2021-03] Use [Time_ns_unix]

val pause_forever : [ `Use_Time_ns_unix ] @@ portable

deprecated [since 2021-03] Use [Time_ns_unix]

include module type of struct include Alternate_sexp end

type nonrec t = Core.Time_ns.t

include Bin_prot.Binable.S__local with type t := t

include Bin_prot.Binable.S_only_functions__local with type t := t

include sig ... end

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : t Bin_prot.Read.vtag_reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_compare_lib.Comparable.S with type t := t

include Ppx_compare_lib.Comparable.S__local with type t := t

include Ppx_compare_lib.Equal.S with type t := t

include Ppx_compare_lib.Equal.S__local with type t := t

include Ppx_hash_lib.Hashable.S_any with type t := t

val hash_fold_t : t Ppx_hash_lib.hash_fold

val hash : t -> Ppx_hash_lib.Std.Hash.hash_value

include Sexplib0.Sexpable.S with type t := t

include Sexplib0.Sexpable.Of_sexp with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

include Sexplib0.Sexpable.Sexp_of with type t := t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t @@ portable

include Base.Comparable.S__local with type t := t

include Base.Comparisons.S__local with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness = Core.Time_ns.Alternate_sexp.comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.T.comparator

val validate_lbound : min:t Core.Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Core.Maybe_bound.t -> t Validate.check

val validate_bound : min:t Core.Maybe_bound.t -> max:t Core.Maybe_bound.t -> t Validate.check

module Replace_polymorphic_compare = Core.Core_private.Time_ns_alternate_sexp.Replace_polymorphic_compare

module Map = Core.Core_private.Time_ns_alternate_sexp.Map

module Set = Core.Core_private.Time_ns_alternate_sexp.Set

include Diffable.S_atomic with type t := t

module Diff = Core.Core_private.Time_ns_alternate_sexp.Diff

val after : t -> Span.t -> t

val sexp_of_t : Core.Time_ns.t -> Sexplib0.Sexp.t