Module Synchronous_time_source.Event

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Core.Invariant.S with type t := t

val invariant : t -> unit

module Option : sig ... end

val at : [> Core.read ] T1.t -> Core.Int63.t -> callback -> t

These are like the corresponding run_* functions, except they return an event that one can later abort.

val after : [> Core.read ] T1.t -> Core.Int63.t -> callback -> t

val at_intervals : [> Core.read ] T1.t -> Core.Int63.t -> callback -> t

module Abort_result : sig ... end

val abort : [> Core.read ] T1.t -> t -> Abort_result.t

abort t aborts the event t, if possible, and returns Ok if the event was aborted, or the reason it could not be aborted. abort returns Currently_happening iff it is called on an event while running that event's callback and the event is not scheduled at intervals.

val abort_exn : [> Core.read ] T1.t -> t -> unit

val abort_if_possible : [> Core.read ] T1.t -> t -> unit

val create : [> Core.read ] T1.t -> callback -> t

create timesource callback creates an event that is not scheduled in timesource's timing wheel but is available to be scheduled using schedule_at and schedule_after.

val schedule_at : [> Core.read ] T1.t -> t -> Core.Int63.t -> unit Core.Or_error.t

schedule_at timesource t time schedules t to fire at time. schedule_at returns Error if t is currently scheduled to run.

val schedule_after : [> Core.read ] T1.t -> t -> Core.Int63.t -> unit Core.Or_error.t

val schedule_at_intervals : [> Core.read ] T1.t -> t -> Core.Int63.t -> unit Core.Or_error.t

val reschedule_at : [> Core.read ] T1.t -> t -> Core.Int63.t -> unit

reschedule_at timesource t time updates t to next fire at time.

For periodic events, reschedule updates the next time to fire, and leaves the interval unchanged. Events rescheduled to a past time will fire at the next advance of timesource.

val reschedule_after : [> Core.read ] T1.t -> t -> Core.Int63.t -> unit

val scheduled_at : t -> Core.Int63.t

scheduled_at returns the time that the event is currently scheduled at