Module Throttle.Sequencer

A sequencer is a throttle that is specialized to only allow one job at a time and to, by default, not continue on error.

type 'a t = ( 'a, [ `sequencer ] ) T2.t

val sexp_of_t : ( 'a -> Sexplib0.Sexp.t ) -> 'a t -> Sexplib0.Sexp.t

val create : ?continue_on_error:bool -> 'a -> 'a t