Module Async_kernel__Async_kernel_config

Settings that globally affect the behavior of Async.

These can be set by setting an environment variable, ASYNC_CONFIG, to a sexp representation of the config. Also, setting ASYNC_CONFIG to an invalid sexp (e.g. the empty string), will cause your program to print to stderr a usage message describing how to configure ASYNC_CONFIG, and exit nonzero. For example, the following shell command should print the usage message:

      ASYNC_CONFIG= foo.exe