Module Async_kernel__Bvar

A Bvar is a synchronization point that allows one to broadcast a value to clients waiting on the broadcast. With a Bvar, one can efficiently notify multiple clients of edge-triggered conditions, repeating as each edge trigger occurs.

Bvar is like an ivar/deferred, except that it is always "empty" and can be repeatedly "filled" (via broadcast).

Another way to view Bvar is as a restriction of Condition that supports only broadcast, not signaling a single waiter. Dropping signal simplifies the implementation significantly.

The 'permissions parameter is used for read/write permissions. Also see Perms.