Module Async_kernel__Lazy_deferred

A delayed computation that can produce a deferred.

Nothing happens with a lazy deferred unless one forces it. Forcing a lazy deferred starts the computation, which will eventually cause the deferred to become determined. As usual with laziness, multiply forcing a lazy deferred is no different than forcing it a single time.

Exceptions (both synchronous and asynchronous) raised by a delayed computation are returned by force (wait, peek, etc.), or will be raised to the monitor in effect when force_exn (wait_exn, peek_exn, etc.) was called.

The type is not exposed nor defined as 'a Deferred.t Lazy.t or 'a Or_error.t Deferred.t Lazy.t, because there is a difference in power with these types. Any value of type 'a Deferred.t Lazy.t would mishandle asynchronous exceptions in the computation of 'a. For instance, the following code blocks forever regardless of how v is defined:

let v : Nothing.t Deferred.t Lazy.t = lazy (return "" >>| failwith) in
let%bind _ = try_with (fun () -> force v) in
let%bind _ = try_with (fun () -> force v) in

There is no val of_lazy : 'a Deferred.t Lazy.t -> 'a t because of the difference in power.

See also Deferred.Memo.unit, if you only are interested in create and force.