Module Async_kernel__Monitor

The part of the Execution_context that determines what to do when there is an unhandled exception.

Every Async computation runs within the context of some monitor, which, when the computation is running, is referred to as the "current" monitor. Monitors are arranged in a tree -- when a new monitor is created, it is a child of the current monitor.

If a computation raises an unhandled exception, the behavior depends on whether the current monitor is "detached" or "attached". If the monitor has been "detached", via one of the detach* functions, then whoever detached it is responsible for dealing with the exception. If the monitor is still attached, then the exception bubbles to monitor's parent. If an exception bubbles to the initial monitor, i.e., the root of the monitor tree, that prints an unhandled-exception message and calls exit 1.

Note about the toplevel monitor

It is important to note that in the toplevel monitor, exceptions will only be caught in the Async part of a computation. For example, in:

upon (f ()) g 

if f raises, the exception will not go to a monitor; it will go to the next caml exception handler on the stack. Any exceptions raised by g will be caught by the scheduler and propagated to the toplevel monitor. Because of this it is advised to always use Scheduler.schedule or Scheduler.within. For example:

Scheduler.within (fun () -> upon (f ()) g) 

This code will catch an exception in either f or g, and propagate it to the monitor.

This is only relevant to the toplevel monitor because if you create another monitor and you wish to run code within it you have no choice but to use Scheduler.within. try_with creates its own monitor and uses Scheduler.within, so it does not have this problem.