Module Async_kernel__Mvar

An Mvar is a mutable location that is either empty or contains a value. One can put or set the value, and wait on value_available for the location to be filled in either way.

Having an Mvar.Read_write.t gives the capability to mutate the mvar.

The key difference between an Mvar and an Ivar is that an Mvar may be filled multiple times.

This implementation of Mvar also allows one to replace the value without any guarantee that the reading side has seen it. This is useful in situations where last-value semantics are desired (e.g., you want to signal whenever a config file is updated, but only care about the most recent contents).

An Mvar can also be used as a baton-passing mechanism between a producer and consumer. For instance, a producer reading from a socket and producing a set of deserialized messages can put the batch from a single read into an Mvar and can wait for taken to return as a pushback mechanism. The consumer meanwhile waits on value_available. This way the natural batch size is passed between the two sub-systems with minimal overhead.