Module Eager_deferred

module type S = sig ... end

include S with type 'a t := 'a Async_kernel.Deferred.t

include Async_kernel.Invariant.S1 with type 'a t := 'a Async_kernel.Deferred.t

val invariant : ('a -> unit) -> 'a Async_kernel.Deferred.t -> unit

include Core_kernel.Monad with type 'a t := 'a Async_kernel.Deferred.t

val (>>=) : 'a Async_kernel.Deferred.t -> ('a -> 'a Async_kernel.Deferred.t) -> 'a Async_kernel.Deferred.t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a Async_kernel.Deferred.t -> ('a -> 'b) -> 'a Async_kernel.Deferred.t

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a Async_kernel.Deferred.t -> f:('a -> 'a Async_kernel.Deferred.t) -> 'a Async_kernel.Deferred.t

bind t ~f = t >>= f

val return : 'a -> 'a Async_kernel.Deferred.t

return v returns the (trivial) computation that returns v.

val map : 'a Async_kernel.Deferred.t -> f:('a -> 'b) -> 'a Async_kernel.Deferred.t

map t ~f is t >>| f.

val join : 'a Async_kernel.Deferred.t -> 'a Async_kernel.Deferred.t

join t is t >>= (fun t' -> t').

val ignore_m : 'a Async_kernel.Deferred.t -> 'a Async_kernel.Deferred.t

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a Async_kernel.Deferred.t list -> 'a Async_kernel.Deferred.t

val all_unit : 'a Async_kernel.Deferred.t list -> 'a Async_kernel.Deferred.t

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end

module Infix : sig ... end

val any : 'a Async_kernel.Deferred.t list -> 'a Async_kernel.Deferred.t

val any_unit : 'a Async_kernel.Deferred.t list -> 'a Async_kernel.Deferred.t

val both : 'a Async_kernel.Deferred.t -> 'a Async_kernel.Deferred.t -> 'a Async_kernel.Deferred.t

val create : ('a Async_kernel.Ivar.t -> unit) -> 'a Async_kernel.Deferred.t

val don't_wait_for : 'a Async_kernel.Deferred.t -> unit

val ignore : 'a Async_kernel.Deferred.t -> 'a Async_kernel.Deferred.t

val is_determined : 'a Async_kernel.Deferred.t -> bool

val never : unit -> 'a Async_kernel.Deferred.t

val ok : 'a Async_kernel.Deferred.t -> 'a Async_kernel.Deferred.t

val peek : 'a Async_kernel.Deferred.t -> 'a option

val unit : 'a Async_kernel.Deferred.t

val upon : 'a Async_kernel.Deferred.t -> ('a -> unit) -> unit

val value_exn : 'a Async_kernel.Deferred.t -> 'a

val repeat_until_finished : 'state -> ('state -> 'a Async_kernel.Deferred.t) -> 'a Async_kernel.Deferred.t

module List : Async_kernel.Monad_sequence.S with type 'a monad := 'a Async_kernel.Deferred.t with type 'a t := 'a list

module Or_error : sig ... end

module Use : sig ... end

Intended usage is to open Eager_deferred.Use to shadow operations from the non-eager world and rebind them to their eager counterparts.