Module Eager_deferred.Or_error

type 'a t = 'a Async_kernel.Deferred.t

include Core_kernel.Monad.S with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'a t) -> 'a t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ('a -> 'b) -> 'a t

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'a t) -> 'a t

bind t ~f = t >>= f

val return : 'a -> 'a t

return v returns the (trivial) computation that returns v.

val map : 'a t -> f:('a -> 'b) -> 'a t

map t ~f is t >>| f.

val join : 'a t -> 'a t

join t is t >>= (fun t' -> t').

val ignore_m : 'a t -> 'a t

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a t list -> 'a t

val all_unit : 'a t list -> 'a t

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end

val fail : Core_kernel.Error.t -> _ t

val ok_unit : unit t

val ignore : _ t -> unit t

val ok_exn : 'a t -> 'a Async_kernel.Deferred.t

val of_exn : exn -> _ t

val of_exn_result : 'a Async_kernel.Deferred.t -> 'a t

val error : string -> 'a -> ('a -> Core_kernel.Sexp.t) -> _ t

val error_s : Core_kernel.Sexp.t -> _ t

val error_string : string -> _ t

val errorf : ('a, unit, string, _ t) Core_kernel.format4 -> 'a

val tag : 'a t -> tag:string -> 'a t

val tag_s : 'a t -> tag:Core_kernel.Sexp.t -> 'a t

val tag_arg : 'a t -> string -> 'b -> ('b -> Core_kernel.Sexp.t) -> 'a t

val unimplemented : string -> _ t

val find_map_ok : 'a list -> f:('a -> 'b t) -> 'b t

val try_with : ?⁠extract_exn:bool -> ?⁠run:[ `Now | `Schedule ] -> ?⁠here:Stdlib.Lexing.position -> ?⁠name:string -> (unit -> 'a Async_kernel.Deferred.t) -> 'a t

Note that try_with f is eager only in the Ok case.

val try_with_join : ?⁠extract_exn:bool -> ?⁠run:[ `Now | `Schedule ] -> ?⁠here:Stdlib.Lexing.position -> ?⁠name:string -> (unit -> 'a t) -> 'a t

Note that try_with_join f is eager only when no exception is raised by f.

val combine_errors : 'a t list -> 'a list t

val combine_errors_unit : unit t list -> unit t

val filter_ok_at_least_one : 'a t list -> 'a list t

module List : Async_kernel.Monad_sequence.S with type 'a monad := 'a t with type 'a t := 'a list

val repeat_until_finished : 'state -> ('state -> [ `Repeat of 'state | `Finished of 'result ] t) -> 'result t