Module type Eager_deferred.S

type +'a t

include Async_kernel.Invariant.S1 with type 'a t := 'a t

val invariant : ('a -> unit) -> 'a t -> unit

include Core_kernel.Monad with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'b t) -> 'b t

bind t ~f = t >>= f

val return : 'a -> 'a t

return v returns the (trivial) computation that returns v.

val map : 'a t -> f:('a -> 'b) -> 'b t

map t ~f is t >>| f.

val join : 'a t t -> 'a t

join t is t >>= (fun t' -> t').

val ignore_m : 'a t -> unit t

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end

These are convenient to have in scope when programming with a monad:

module Infix : sig ... end

val any : 'a t list -> 'a t

val any_unit : unit t list -> unit t

val both : 'a t -> 'b t -> ('a * 'b) t

val create : ('a Async_kernel.Ivar.t -> unit) -> 'a t

val don't_wait_for : unit t -> unit

val is_determined : 'a t -> bool

val never : unit -> _ t

val ok : 'a t -> ('a, _) Core_kernel.Result.t t

val peek : 'a t -> 'a option

val unit : unit t

val upon : 'a t -> ('a -> unit) -> unit

val value_exn : 'a t -> 'a

val repeat_until_finished : 'state -> ('state -> [ `Repeat of 'state | `Finished of 'result ] t) -> 'result t

module List : Async_kernel.Monad_sequence.S with type 'a monad := 'a t with type 'a t := 'a list

module Or_error : sig ... end

module Use : sig ... end

Intended usage is to open Eager_deferred.Use to shadow operations from the non-eager world and rebind them to their eager counterparts.