Module S.Use

Intended usage is to open Eager_deferred.Use to shadow operations from the non-eager world and rebind them to their eager counterparts.

module Deferred : sig ... end

include Core_kernel.Monad.Infix with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'a t) -> 'a t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ('a -> 'b) -> 'a t

t >>| f is t >>= (fun a -> return (f a)).

include module type of Deferred.Let_syntax

val return : 'a -> 'a Deferred.t

val (>>=) : 'a Deferred.t -> ('a -> 'a Deferred.t) -> 'a Deferred.t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a Deferred.t -> ('a -> 'b) -> 'a Deferred.t

t >>| f is t >>= (fun a -> return (f a)).

module Let_syntax : sig ... end

val upon : 'a t -> ('a -> unit) -> unit

val (>>>) : 'a t -> ('a -> unit) -> unit

val (>>=?) : ('a, 'e) Core_kernel.Result.t t -> ('a -> ('b, 'e) Core_kernel.Result.t t) -> ('b, 'e) Core_kernel.Result.t t

val (>>|?) : ('a, 'e) Core_kernel.Result.t t -> ('a -> 'b) -> ('b, 'e) Core_kernel.Result.t t