Module Deferred.Infix

include Core_kernel.Monad.Infix with type 'a t := 'a t

val (>>=) : 'a t -> ('a -> 'a t) -> 'a t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ('a -> 'b) -> 'a t

t >>| f is t >>= (fun a -> return (f a)).

val (>>>) : 'a t -> ('a -> unit) -> unit