Module Persistent_connection_kernel__Persistent_connection_kernel_intf

An actively maintained connection to some service that eagerly and repeatedly attempts to reconnect whenever the underlying connection is lost, until a new one can be established.