Module Gc.Alarm

A GC alarm calls a user function after the end of each major GC cycle.

type t

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val create : (unit -> unit) -> t

create f arranges for f to be called after the end of each major GC cycle, starting with the current cycle or the next one. f will be run in the monitor that create was called in.

val delete : t -> unit

delete t will stop the calls to the function associated with t. Calling delete t again has no effect.