Module Gc.Allocation_policy

The policy used for allocating in the heap.

The Next_fit policy is quite fast but can result in fragmentation.

The First_fit policy can be slower in some cases but can be better for programs with fragmentation problems.

The Best_fit policy is as fast as Next_fit and has less fragmentation than First_fit.

The default is Next_fit.

type t =

| Next_fit

| First_fit

| Best_fit

val compare : t -> t -> Base.Int.t

val equal : t -> t -> Base.Bool.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t