Module Synchronous_time_source.Time_ns

include Core_kernel.Core_kernel_private.Time_ns_alternate_sexp

include Core_kernel.Time_ns

module Span : sig ... end

module Ofday : sig ... end

val hash_fold_t : Base.Hash.state -> Core_kernel.Int63.t -> Base.Hash.state

val hash : Core_kernel.Int63.t -> Base.Hash.hash_value

include Typerep_lib.Typerepable.S with type t := Core_kernel.Int63.t

val typerep_of_t : Core_kernel.Int63.t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : Core_kernel.Int63.t Typerep_lib.Typename.t

include Bin_prot.Binable.S with type t := Core_kernel.Int63.t

include Bin_prot.Binable.S_only_functions with type t := Core_kernel.Int63.t

val bin_size_t : Core_kernel.Int63.t Bin_prot.Size.sizer

val bin_write_t : Core_kernel.Int63.t Bin_prot.Write.writer

val bin_read_t : Core_kernel.Int63.t Bin_prot.Read.reader

val __bin_read_t__ : (int -> Core_kernel.Int63.t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : Core_kernel.Int63.t Bin_prot.Type_class.writer

val bin_reader_t : Core_kernel.Int63.t Bin_prot.Type_class.reader

val bin_t : Core_kernel.Int63.t Bin_prot.Type_class.t

include Base.Comparisons.S with type t := Core_kernel.Int63.t

include Base.Comparisons.Infix with type t := Core_kernel.Int63.t

val (>=) : Core_kernel.Int63.t -> Core_kernel.Int63.t -> bool

val (<=) : Core_kernel.Int63.t -> Core_kernel.Int63.t -> bool

val (=) : Core_kernel.Int63.t -> Core_kernel.Int63.t -> bool

val (>) : Core_kernel.Int63.t -> Core_kernel.Int63.t -> bool

val (<) : Core_kernel.Int63.t -> Core_kernel.Int63.t -> bool

val (<>) : Core_kernel.Int63.t -> Core_kernel.Int63.t -> bool

val equal : Core_kernel.Int63.t -> Core_kernel.Int63.t -> bool

val compare : Core_kernel.Int63.t -> Core_kernel.Int63.t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : Core_kernel.Int63.t -> Core_kernel.Int63.t -> Core_kernel.Int63.t

val max : Core_kernel.Int63.t -> Core_kernel.Int63.t -> Core_kernel.Int63.t

module Alternate_sexp : sig ... end

Note that we expose a sexp format that is not the one exposed in Core.

include Core_kernel.Quickcheck.S_range with type t := Core_kernel.Int63.t

include Core_kernel.Quickcheck_intf.S with type t := Core_kernel.Int63.t

val quickcheck_generator : Core_kernel.Int63.t Base_quickcheck.Generator.t

val quickcheck_observer : Core_kernel.Int63.t Base_quickcheck.Observer.t

val quickcheck_shrinker : Core_kernel.Int63.t Base_quickcheck.Shrinker.t

val gen_incl : Core_kernel.Int63.t -> Core_kernel.Int63.t -> Core_kernel.Int63.t Base_quickcheck.Generator.t

gen_incl lower_bound upper_bound produces values between lower_bound and upper_bound, inclusive. It uses an ad hoc distribution that stresses boundary conditions more often than a uniform distribution, while still able to produce any value in the range. Raises if lower_bound > upper_bound.

val gen_uniform_incl : Core_kernel.Int63.t -> Core_kernel.Int63.t -> Core_kernel.Int63.t Base_quickcheck.Generator.t

gen_uniform_incl lower_bound upper_bound produces a generator for values uniformly distributed between lower_bound and upper_bound, inclusive. Raises if lower_bound > upper_bound.

Comparisons

val is_earlier : Core_kernel.Int63.t -> than:Core_kernel.Int63.t -> Base.Bool.t

val is_later : Core_kernel.Int63.t -> than:Core_kernel.Int63.t -> Base.Bool.t

Conversions

val of_date_ofday : zone:Core_kernel__Zone.t -> Core_kernel__Date0.t -> Ofday.t -> Core_kernel.Int63.t

val of_date_ofday_precise : Core_kernel__Date0.t -> Ofday.t -> zone:Core_kernel__Zone.t -> [ `Once of Core_kernel.Int63.t | `Twice of Core_kernel.Int63.t * Core_kernel.Int63.t | `Never of Core_kernel.Int63.t ]

Because timezone offsets change throughout the year (clocks go forward or back) some local times can occur twice or not at all. In the case that they occur twice, this function gives `Twice with both occurrences in order; if they do not occur at all, this function gives `Never with the time at which the local clock skips over the desired time of day.

Note that this is really only intended to work with DST transitions and not unusual or dramatic changes, like the calendar change in 1752 (run "cal 9 1752" in a shell to see). In particular it makes the assumption that midnight of each day is unambiguous.

Most callers should use of_date_ofday rather than this function. In the `Twice and `Never cases, of_date_ofday will return reasonable times for most uses.

val to_date_ofday : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Core_kernel__Date0.t * Ofday.t

val to_date_ofday_precise : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Core_kernel__Date0.t * Ofday.t * [ `Only | `Also_at of Core_kernel.Int63.t | `Also_skipped of Core_kernel__Date0.t * Ofday.t ]

Always returns the Date.t * Ofday.t that to_date_ofday would have returned, and in addition returns a variant indicating whether the time is associated with a time zone transition.

      - `Only         -> there is a one-to-one mapping between [t]'s and
                         [Date.t * Ofday.t] pairs
      - `Also_at      -> there is another [t] that maps to the same [Date.t * Ofday.t]
                         (this date/time pair happened twice because the clock fell back)
      - `Also_skipped -> there is another [Date.t * Ofday.t] pair that never happened (due
                         to a jump forward) that [of_date_ofday] would map to the same
                         [t].

val to_date : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Core_kernel__Date0.t

val to_ofday : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Ofday.t

val reset_date_cache : Base.Unit.t -> Base.Unit.t

For performance testing only; reset_date_cache () resets an internal cache used to speed up to_date and related functions when called repeatedly on times that fall within the same day.

val epoch : Core_kernel.Int63.t

midnight, Jan 1, 1970 in UTC

val convert : from_tz:Core_kernel__Zone.t -> to_tz:Core_kernel__Zone.t -> Core_kernel__Date0.t -> Ofday.t -> Core_kernel__Date0.t * Ofday.t

It's unspecified what happens if the given date/ofday/zone correspond to more than one date/ofday pair in the other zone.

val utc_offset : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Span.t

Other string conversions

The {to,of}_string functions in Time convert to UTC time, because a local time zone is not necessarily available. They are generous in what they will read in.

val of_string : string -> Core_kernel.Int63.t

val to_string : Core_kernel.Int63.t -> string

val to_filename_string : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Base.String.t

to_filename_string t ~zone converts t to string with format YYYY-MM-DD_HH-MM-SS.mmm which is suitable for using in filenames.

val of_filename_string : Base.String.t -> zone:Core_kernel__Zone.t -> Core_kernel.Int63.t

of_filename_string s ~zone converts s that has format YYYY-MM-DD_HH-MM-SS.mmm into time.

val to_string_abs : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Base.String.t

to_string_abs ~zone t is the same as to_string t except that it uses the given time zone.

val to_string_abs_trimmed : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Base.String.t

to_string_abs_trimmed is the same as to_string_abs, but drops trailing seconds and milliseconds if they are 0.

val to_string_abs_parts : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Base.String.t Base.List.t

val to_string_trimmed : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Base.String.t

Same as to_string_abs_trimmed, except it leaves off the timezone, so won't reliably round trip.

val to_sec_string : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Base.String.t

Same as to_string_abs, but without milliseconds

val of_localized_string : zone:Core_kernel__Zone.t -> Base.String.t -> Core_kernel.Int63.t

of_localized_string ~zone str read in the given string assuming that it represents a time in zone and return the appropriate Time.t

val of_string_gen : default_zone:(Base.Unit.t -> Core_kernel__Zone.t) -> find_zone:(Base.String.t -> Core_kernel__Zone.t) -> Base.String.t -> Core_kernel.Int63.t

of_string_gen ~default_zone ~find_zone s attempts to parse s as a t, calling out to default_zone and find_zone as needed.

val to_string_iso8601_basic : Core_kernel.Int63.t -> zone:Core_kernel__Zone.t -> Base.String.t

to_string_iso8601_basic return a string representation of the following form: %Y-%m-%dT%H:%M:%S.%s%Z e.g. to_string_iso8601_basic ~zone:Time.Zone.utc epoch = "1970-01-01T00:00:00.000000Z"

val occurrence : [ `First_after_or_at | `Last_before_or_at ] -> Core_kernel.Int63.t -> ofday:Ofday.t -> zone:Core_kernel__Zone.t -> Core_kernel.Int63.t

occurrence side time ~ofday ~zone returns a Time.t that is the occurrence of ofday (in the given zone) that is the latest occurrence (<=) time or the earliest occurrence (>=) time, according to side.

NOTE: If the given time converted to wall clock time in the given zone is equal to ofday then the t returned will be equal to the t given.

val epoch : Core_kernel.Int63.t

Unix epoch (1970-01-01 00:00:00 UTC)

val min_value_representable : Core_kernel.Int63.t

The minimum representable time.

val max_value_representable : Core_kernel.Int63.t

The maximum representable time.

val min_value_for_1us_rounding : Core_kernel.Int63.t

The minimum time that rounds to a Time.t with microsecond precision.

val max_value_for_1us_rounding : Core_kernel.Int63.t

The maximum time that rounds to a Time.t with microsecond precision.

val min_value : Core_kernel.Int63.t

An alias for min_value_for_1us_rounding.

val max_value : Core_kernel.Int63.t

An alias for max_value_for_1us_rounding.

val now : Base.Unit.t -> Core_kernel.Int63.t

The current time.

val add : Core_kernel.Int63.t -> Span.t -> Core_kernel.Int63.t

overflows silently

val add_saturating : Core_kernel.Int63.t -> Span.t -> Core_kernel.Int63.t

As add; rather than over/underflowing, clamps the result to the closed interval between min_value_representable and max_value_representable.

val sub_saturating : Core_kernel.Int63.t -> Span.t -> Core_kernel.Int63.t

As sub; rather than over/underflowing, clamps the result to the closed interval between min_value_representable and max_value_representable.

val sub : Core_kernel.Int63.t -> Span.t -> Core_kernel.Int63.t

overflows silently

val next : Core_kernel.Int63.t -> Core_kernel.Int63.t

overflows silently

val prev : Core_kernel.Int63.t -> Core_kernel.Int63.t

overflows silently

val diff : Core_kernel.Int63.t -> Core_kernel.Int63.t -> Span.t

overflows silently

val abs_diff : Core_kernel.Int63.t -> Core_kernel.Int63.t -> Span.t

overflows silently

val to_span_since_epoch : Core_kernel.Int63.t -> Span.t

val of_span_since_epoch : Span.t -> Core_kernel.Int63.t

val to_int63_ns_since_epoch : Core_kernel.Int63.t -> Core_kernel.Int63.t

val of_int63_ns_since_epoch : Core_kernel.Int63.t -> Core_kernel.Int63.t

val to_int_ns_since_epoch : Core_kernel.Int63.t -> Base.Int.t

Will raise on 32-bit platforms. Consider to_int63_ns_since_epoch instead.

val of_int_ns_since_epoch : Base.Int.t -> Core_kernel.Int63.t

val next_multiple : ?⁠can_equal_after:Base.Bool.t -> base:Core_kernel.Int63.t -> after:Core_kernel.Int63.t -> interval:Span.t -> Base.Unit.t -> Core_kernel.Int63.t

next_multiple ~base ~after ~interval returns the smallest time of the form:

time = base + k * interval

where k >= 0 and time > after. It is an error if interval <= 0.

Supplying ~can_equal_after:true allows the result to satisfy time >= after.

Overflows silently.

val prev_multiple : ?⁠can_equal_before:Base.Bool.t -> base:Core_kernel.Int63.t -> before:Core_kernel.Int63.t -> interval:Span.t -> Base.Unit.t -> Core_kernel.Int63.t

prev_multiple ~base ~before ~interval returns the largest time of the form:

time = base + k * interval

where k >= 0 and time < before. It is an error if interval <= 0.

Supplying ~can_equal_before:true allows the result to satisfy time <= before.

val random : ?⁠state:Base.Random.State.t -> Base.Unit.t -> Core_kernel.Int63.t

val of_time : Base.Float.t -> Core_kernel.Int63.t

val to_time : Core_kernel.Int63.t -> Base.Float.t

val to_time_float_round_nearest : Core_kernel.Int63.t -> Base.Float.t

val to_time_float_round_nearest_microsecond : Core_kernel.Int63.t -> Base.Float.t

val of_time_float_round_nearest : Base.Float.t -> Core_kernel.Int63.t

val of_time_float_round_nearest_microsecond : Base.Float.t -> Core_kernel.Int63.t

module Utc : sig ... end

module Stable : sig ... end

include Alternate_sexp

type nonrec t = Core_kernel.Int63.t

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val after : Core_kernel.Int63.t -> Span.t -> Core_kernel.Int63.t

val format : float -> string -> string

val sexp_of_t : Core_kernel.Int63.t -> Ppx_sexp_conv_lib.Sexp.t