Module Limiter_async.Expert

val kill : t -> unit

kills t, which aborts all enqueued jobs that haven't started and all jobs enqueued in the future. If t has already been killed, then calling kill t has no effect. Note that kill does not affect currently running jobs in any way.

val is_dead : t -> bool

is_dead t returns true if t was killed, either by kill or by an unhandled exception in a job.

val cost_of_jobs_waiting_to_start : t -> int

returns the total cost of all jobs that have been enqueued but have not yet started.

val to_jane_limiter : t -> Limiter.t

returns the underlying limiter. It is an error to do anything with the limiter that isn't a read-only operation.